Task 3.1.3 REST API

+++ src/main/java/ru/itmentor/spring/boot_security/demo/service/UserServiceImpl.java
@@ -6,22 +6,29 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.itmentor.spring.boot_security.demo.exception.users_exceptions.UserAlreadyExistsException;
+import ru.itmentor.spring.boot_security.demo.exception.users_exceptions.UserNotFoundException;
+import ru.itmentor.spring.boot_security.demo.model.Role;
 import ru.itmentor.spring.boot_security.demo.model.User;
 import ru.itmentor.spring.boot_security.demo.repositories.UserRepository;
 import ru.itmentor.spring.boot_security.demo.security.UserDetailsWrapper;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 @Service
 public class UserServiceImpl implements UserService {
 
-
+    private final RoleService roleService;
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
 
+
     @Autowired
-    public UserServiceImpl(UserRepository userRepository, PasswordEncoder passwordEncoder) {
+    public UserServiceImpl(RoleService roleService, UserRepository userRepository, PasswordEncoder passwordEncoder) {
+        this.roleService = roleService;
         this.userRepository = userRepository;
         this.passwordEncoder = passwordEncoder;
     }
@@ -35,28 +42,49 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @Override
     public void saveUser(User user) {
+        if (userRepository.findByUsername(user.getUsername()).isPresent()) {
+            throw new UserAlreadyExistsException();
+        }
+        checkRoles(user);
         user.setPassword(passwordEncoder.encode(user.getPassword()));
         userRepository.save(user);
     }
 
     @Transactional
     @Override
-    public void updateUser(User user) {
-        if(user.getPassword().equals(getUserById(user.getId()).getPassword())){userRepository.save(user);}else {
-            user.setPassword(passwordEncoder.encode(user.getPassword()));
-            userRepository.save(user);
+    public void updateUser(User user, Long id) {
+        if (userRepository.findById(id).isEmpty()) {
+            throw new UserNotFoundException(id);
         }
+        user.setId(id);
+        checkRoles(user);
+        if (!user.getPassword().equals(getUserById(user.getId()).getPassword())) {
+            user.setPassword(passwordEncoder.encode(user.getPassword()));
+        }
+        userRepository.save(user);
+    }
+
+    private void checkRoles(User user) {
+        Set<Role> roleSet = new HashSet<>();
+        for (Role roleUser : user.getRoles()) {
+            Optional<Role> existingRole = roleService.findByRoleName(roleUser.getRoleName());
+            existingRole.ifPresent(roleSet::add);
+        }
+        user.setRoles(roleSet);
     }
 
     @Transactional(readOnly = true)
     @Override
     public User getUserById(Long id) {
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id).orElseThrow(UserNotFoundException::new);
     }
 
     @Transactional
     @Override
     public void deleteUser(Long id) {
+        if (userRepository.findById(id).isEmpty()) {
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
     }
